pyproject.toml
[project]
name = "tts-rack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cattrs>=25.3.0",
    "dsp-board",
    "einops>=0.8.1",
    "fastdtw>=0.3.4",
    "hydra-core>=1.3.2",
    "jaconv>=0.4.0",
    "lightning>=2.5.1",
    "matplotlib>=3.10.7",
    "metric-board",
    "monotonic-align",
    "pathlib>=1.0.1",
    "pydomino",
    "pyopenjtalk>=0.4.1",
    "questionary>=2.1.1",
    "soundfile>=0.13.1",
    "torch>=2.5.1",
    "torchaudio>=2.5.1, <2.8.0",
    "tqdm>=4.67.1",
    "wandb>=0.21.4",
]

[tool.uv.sources]
pydomino = { git = "https://github.com/DwangoMediaVillage/pydomino" }
monotonic-align = { git = "https://github.com/resemble-ai/monotonic_align.git" }
dsp-board = { git = "https://github.com/STnvc7/dsp_board.git" }
metric-board = { git = "https://github.com/STnvc7/metric_board.git" }
